Added changes() method.
This allows being able to get a changefeed for the table. The Type def for the calllback are incorrect due to it not returning the model but a "document" but we don't have a clean way to return the definition of the raw model properties.
Also removed an unneeded reference for types.