Changed: Allow the setting of the table name.
Currently if you have more than one model named the same on one connection it will use the last added one. Changing the class name fixes this but then you have more tables created with the wrong names. Thus that this setting allows the option to set the table name that should be used.